[New] H2 Db Support + improvements (#774) by Fuzzlemann
* Adding H2 Database Support

* Making HtmlCompressor a constant so it is not created new each time it is used.
Note: It is thread-safe (see: https://code.google.com/archive/p/htmlcompressor/ "HtmlCompressor and XmlCompressor classes are considered thread safe* and can be used in multi-thread environment")

* Adds the H2 Database to the used libraries in the README.md

* Changes all invalid license headers

* Refactors many methods into an enum named DBType
These methods are:
 -> getName()
 -> getConfigName()
 -> supportsMySQLQueries()

 to a check if the DBType is H2

* KeepAliveTask splitted into separate class to remove duplicate code

* Fixes compilation error

* Refactors Database Tests into one Common class with all generic Database Tests and specific Database Tests

* Fixes some JavaDocs

* Adds license headers

* Optimizing comparisons to enum comparisons
