+++ Plan/pom.xml
@@ -171,6 +171,13 @@
             <version>1.5.2</version>
         </dependency>
 
+        <!-- H2 -->
+        <dependency>
+            <groupId>com.h2database</groupId>
+            <artifactId>h2</artifactId>
+            <version>1.4.196</version> <!-- This version is necessary so Sponge will not crash -->
+        </dependency>
+
         <!-- Mockito (Test Dependency) -->
         <dependency>
             <groupId>org.mockito</groupId>
@@ -179,6 +186,7 @@
             <type>jar</type>
             <scope>test</scope>
         </dependency>
+
         <!-- SQLite (Test Dependency) -->
         <dependency>
             <groupId>org.xerial</groupId>
@@ -256,7 +264,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
-                <version>3.1.1</version>
+                <version>3.2.0</version>
                 <executions>
                     <execution>
                         <phase>package</phase>
@@ -301,6 +309,10 @@
                         <relocation>
                             <pattern>com.zaxxer</pattern>
                             <shadedPattern>plan.com.zaxxer</shadedPattern>
+                        </relocation>
+                        <relocation>
+                            <pattern>org.h2</pattern>
+                            <shadedPattern>plan.org.h2</shadedPattern>
                         </relocation>
                         <!--<relocation>-->
                         <!--<pattern>org.slf4j</pattern>-->
