+++ Plan/src/main/java/com/djrapitops/plan/command/commands/InfoCommand.java
@@ -76,7 +76,7 @@ public class InfoCommand extends CommandNode {
                 "",
                 locale.getString(CommandLang.INFO_VERSION, plugin.getVersion()),
                 locale.getString(CommandLang.INFO_UPDATE, updateAvailable),
-                locale.getString(CommandLang.INFO_DATABASE, dbSystem.getDatabase().getName()),
+                locale.getString(CommandLang.INFO_DATABASE, dbSystem.getDatabase().getType().getName()),
                 locale.getString(CommandLang.INFO_BUNGEE_CONNECTION, connectedToBungee),
                 "",
                 ">"
