+++ Plan/src/main/java/com/djrapitops/plan/system/database/BukkitDBSystem.java
@@ -17,6 +17,7 @@
 package com.djrapitops.plan.system.database;
 
 import com.djrapitops.plan.api.exceptions.EnableException;
+import com.djrapitops.plan.system.database.databases.sql.H2DB;
 import com.djrapitops.plan.system.database.databases.sql.MySQLDB;
 import com.djrapitops.plan.system.database.databases.sql.SQLiteDB;
 import com.djrapitops.plan.system.locale.Locale;
@@ -44,15 +45,17 @@ public class BukkitDBSystem extends DBSystem {
             Locale locale,
             MySQLDB mySQLDB,
             SQLiteDB.Factory sqLiteDB,
+            H2DB.Factory h2DB,
             PlanConfig config,
             PluginLogger logger,
             Timings timings,
             ErrorHandler errorHandler
     ) {
-        super(locale, sqLiteDB, logger, timings, errorHandler);
+        super(locale, sqLiteDB, h2DB, logger, timings, errorHandler);
         this.config = config;
 
         databases.add(mySQLDB);
+        databases.add(h2DB.usingDefaultFile());
         databases.add(sqLiteDB.usingDefaultFile());
     }
