+++ Plan/src/main/java/com/djrapitops/plan/system/database/databases/Database.java
@@ -47,18 +47,13 @@ public abstract class Database {
 
     public abstract SaveOperations save();
 
-    public abstract String getName();
-
     /**
-     * Used to get the config name of the database type.
-     * <p>
-     * Thread safe.
+     * Used to get the {@code DBType} of the Database
      *
-     * @return sqlite/mysql
+     * @return the {@code DBType}
+     * @see DBType
      */
-    public String getConfigName() {
-        return getName().toLowerCase().trim();
-    }
+    public abstract DBType getType();
 
     public abstract void close() throws DBException;
