+++ Plan/src/main/java/com/djrapitops/plan/system/database/databases/sql/SQLDB.java
@@ -419,10 +419,6 @@ public abstract class SQLDB extends Database {
         return pingTable;
     }
 
-    public boolean isUsingMySQL() {
-        return false;
-    }
-
     @Override
     public BackupOperations backup() {
         return backupOps;
@@ -468,12 +464,12 @@ public abstract class SQLDB extends Database {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         SQLDB sqldb = (SQLDB) o;
-        return getName().equals(sqldb.getName());
+        return getType() == sqldb.getType();
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getName());
+        return Objects.hash(getType().getName());
     }
 
     public Supplier<UUID> getServerUUIDSupplier() {
