+++ Plan/src/main/java/com/djrapitops/plan/system/database/databases/sql/patches/Patch.java
@@ -16,6 +16,7 @@
  */
 package com.djrapitops.plan.system.database.databases.sql.patches;
 
+import com.djrapitops.plan.system.database.databases.DBType;
 import com.djrapitops.plan.system.database.databases.sql.SQLDB;
 import com.djrapitops.plan.system.database.databases.sql.processing.QueryAllStatement;
 import com.djrapitops.plan.system.database.databases.sql.processing.QueryStatement;
@@ -30,11 +31,11 @@ import java.util.UUID;
 public abstract class Patch {
 
     protected final SQLDB db;
-    protected final boolean usingMySQL;
+    protected final DBType dbType;
 
     public Patch(SQLDB db) {
         this.db = db;
-        usingMySQL = db.isUsingMySQL();
+        this.dbType = db.getType();
     }
 
     public abstract boolean hasBeenApplied();
@@ -46,15 +47,25 @@ public abstract class Patch {
     }
 
     public boolean hasTable(String tableName) {
-        String sql = usingMySQL ?
-                "SELECT * FROM information_schema.TABLES WHERE table_name=? AND TABLE_SCHEMA=? LIMIT 1" :
-                "SELECT tbl_name FROM sqlite_master WHERE tbl_name=?";
+        boolean secondParameter;
+
+        String sql;
+        if (dbType == DBType.H2) {
+            sql = "SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME=?";
+            secondParameter = false;
+        } else if (dbType.supportsMySQLQueries()) {
+            sql = "SELECT * FROM information_schema.TABLES WHERE table_name=? AND TABLE_SCHEMA=? LIMIT 1";
+            secondParameter = true;
+        } else {
+            sql = "SELECT tbl_name FROM sqlite_master WHERE tbl_name=?";
+            secondParameter = false;
+        }
 
         return query(new QueryStatement<Boolean>(sql) {
             @Override
             public void prepare(PreparedStatement statement) throws SQLException {
                 statement.setString(1, tableName);
-                if (usingMySQL) {
+                if (secondParameter) {
                     statement.setString(2, db.getConfig().getString(Settings.DB_DATABASE));
                 }
             }
@@ -67,36 +78,49 @@ public abstract class Patch {
     }
 
     protected boolean hasColumn(String tableName, String columnName) {
-        return usingMySQL ?
-                query(new QueryStatement<Boolean>("SELECT * FROM information_schema.COLUMNS" +
-                        " WHERE TABLE_NAME=? AND COLUMN_NAME=? AND TABLE_SCHEMA=?") {
-                    @Override
-                    public void prepare(PreparedStatement statement) throws SQLException {
-                        statement.setString(1, tableName);
-                        statement.setString(2, columnName);
+        if (dbType.supportsMySQLQueries()) {
+            String query;
+
+            if (dbType == DBType.H2) {
+                query = "SELECT * FROM INFORMATION_SCHEMA.COLUMNS" +
+                        " WHERE TABLE_NAME=? AND COLUMN_NAME=?";
+            } else {
+                query = "SELECT * FROM information_schema.COLUMNS" +
+                        " WHERE TABLE_NAME=? AND COLUMN_NAME=? AND TABLE_SCHEMA=?";
+            }
+
+            return query(new QueryStatement<Boolean>(query) {
+                @Override
+                public void prepare(PreparedStatement statement) throws SQLException {
+                    statement.setString(1, tableName);
+                    statement.setString(2, columnName);
+                    if (dbType != DBType.H2) {
                         statement.setString(3, db.getConfig().getString(Settings.DB_DATABASE));
                     }
+                }
 
-                    @Override
-                    public Boolean processResults(ResultSet set) throws SQLException {
-                        return set.next();
-                    }
-                }) :
-                query(new QueryAllStatement<Boolean>("PRAGMA table_info(" + tableName + ")") {
-                    @Override
-                    public Boolean processResults(ResultSet set) throws SQLException {
-                        while (set.next()) {
-                            if (columnName.equals(set.getString("name"))) {
-                                return true;
-                            }
+                @Override
+                public Boolean processResults(ResultSet set) throws SQLException {
+                    return set.next();
+                }
+            });
+        } else {
+            return query(new QueryAllStatement<Boolean>("PRAGMA table_info(" + tableName + ")") {
+                @Override
+                public Boolean processResults(ResultSet set) throws SQLException {
+                    while (set.next()) {
+                        if (columnName.equals(set.getString("name"))) {
+                            return true;
                         }
-                        return false;
                     }
-                });
+                    return false;
+                }
+            });
+        }
     }
 
     protected void addColumn(String tableName, String columnInfo) {
-        db.executeUnsafe("ALTER TABLE " + tableName + " ADD " + (usingMySQL ? "" : "COLUMN ") + columnInfo);
+        db.executeUnsafe("ALTER TABLE " + tableName + " ADD " + (dbType.supportsMySQLQueries() ? "" : "COLUMN ") + columnInfo);
     }
 
     protected void dropTable(String name) {
@@ -104,7 +128,7 @@ public abstract class Patch {
     }
 
     protected void renameTable(String from, String to) {
-        String sql = usingMySQL ?
+        String sql = dbType.supportsMySQLQueries() ?
                 "RENAME TABLE " + from + " TO " + to :
                 "ALTER TABLE " + from + " RENAME TO " + to;
         db.execute(sql);
