+++ Plan/src/main/java/com/djrapitops/plan/system/database/databases/sql/tables/UsersTable.java
@@ -59,12 +59,12 @@ public class UsersTable extends UserIDTable {
     @Override
     public void createTable() throws DBInitException {
         createTable(TableSqlParser.createTable(tableName)
-                .primaryKeyIDColumn(usingMySQL, Col.ID)
+                .primaryKeyIDColumn(supportsMySQLQueries, Col.ID)
                 .column(Col.UUID, Sql.varchar(36)).notNull().unique()
                 .column(Col.REGISTERED, Sql.LONG).notNull()
                 .column(Col.USER_NAME, Sql.varchar(16)).notNull()
                 .column(Col.TIMES_KICKED, Sql.INT).notNull().defaultValue("0")
-                .primaryKey(usingMySQL, Col.ID)
+                .primaryKey(supportsMySQLQueries, Col.ID)
                 .toString()
         );
     }
@@ -261,13 +261,13 @@ public class UsersTable extends UserIDTable {
      * @return a list of distinct names.
      */
     public List<String> getMatchingNames(String name) {
-        String searchString = "%" + name.toLowerCase() + "%";
+        String searchString = "%" + name + "%";
         NicknamesTable nicknamesTable = db.getNicknamesTable();
         String sql = "SELECT DISTINCT " + Col.USER_NAME + " FROM " + tableName +
-                " WHERE " + Col.USER_NAME + " LIKE ?" +
+                " WHERE LOWER(" + Col.USER_NAME + ") LIKE LOWER(?)" +
                 " UNION SELECT DISTINCT " + Col.USER_NAME + " FROM " + tableName +
                 " INNER JOIN " + nicknamesTable + " on " + Col.ID + "=" + nicknamesTable + "." + NicknamesTable.Col.USER_ID +
-                " WHERE " + NicknamesTable.Col.NICKNAME + " LIKE ?";
+                " WHERE LOWER(" + NicknamesTable.Col.NICKNAME + ") LIKE LOWER(?)";
 
         return query(new QueryStatement<List<String>>(sql, 5000) {
             @Override
