+++ Plan/src/test/java/com/djrapitops/plan/system/database/databases/H2Test.java
@@ -14,25 +14,43 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
  */
-package com.djrapitops.plan.utilities.html;
+package com.djrapitops.plan.system.database.databases;
 
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Test;
-import utilities.RandomData;
+import org.junit.runner.RunWith;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import static org.junit.Assert.assertEquals;
 
 /**
- * @author Rsl1122
+ * Test for the H2 database
+ *
+ * @author Rsl1122, Fuzzlemann
+ * @see SQLiteTest
+ * @since 4.5.1
  */
-public class HtmlUtilsTest {
+@RunWith(MockitoJUnitRunner.Silent.class)
+public class H2Test extends CommonDBTest {
+
+    @BeforeClass
+    public static void setUpClass() throws Exception {
+        handleSetup("H2");
+    }
+
+    @AfterClass
+    public static void tearDownClass() {
+        system.disable();
+    }
 
     @Test
-    public void testRemoveXSS() {
-        String randomString = RandomData.randomString(10);
+    public void testH2GetConfigName() {
+        assertEquals("h2", db.getType().getConfigName());
+    }
 
-        String xss = "<script>" + randomString + "</script><!---->";
-        String result = HtmlUtils.removeXSS(xss);
-
-        assertEquals(randomString, result);
+    @Test
+    public void testH2GetName() {
+        assertEquals("H2", db.getType().getName());
     }
 }
